nuevo
nuevo

+++ usuario/eliminaUser.php
@@ -6,25 +6,29 @@ $user = new User();
 
 if(isset($_SESSION['user'])){
     $user->setUser($userSession->getCurrentUser());
-    $cons="Modifico";
-    $secionusuariosA = $secionusuarios ->get_usuarioDef();
+    $tipo_usuario = $user->getTipoUsuario();
+    $usuarioPri= $tipo_usuario['tipo_usuario'];
+    if ($usuarioPri=="administrador"){
+        $cons="Modifico";
+        $secionusuariosA = $secionusuarios ->get_usuarioDef();
 
-    try {
-        $db = new DB();
-        $pdo = $db->connect();
-        $agregarUserSesion=" DELETE FROM usuario           
-        WHERE UserName='".$secionusuariosA['secionUsr']."'";
+        try {
+            $db = new DB();
+            $pdo = $db->connect();
+            $agregarUserSesion=" DELETE FROM usuario           
+            WHERE UserName='".$secionusuariosA['secionUsr']."'";
 
-        $stmt = $pdo->prepare($agregarUserSesion);
-        $stmt->execute();
-        echo "<script> alert('se a Elimino con exito'); window.location='/base-de-datos-escuela/usuario/mostraraUsuario.php'</script>";
-    } catch (PDOException $e) {
-            echo "Ocurrió un error en la consulta: " . $e->getMessage();
+            $stmt = $pdo->prepare($agregarUserSesion);
+            $stmt->execute();
+            echo "<script> alert('se a Elimino con exito'); window.location='/base-de-datos-escuela/usuario/mostraraUsuario.php'</script>";
+        } catch (PDOException $e) {
+                echo "Ocurrió un error en la consulta: " . $e->getMessage();
+        }
+
+    }else{
+        echo "<script>alert('no cuentas con los permisos para esta opción');window.location='/base-de-datos-escuela/inicio.php'</script>";
     }
-
-
 }else{
-    //echo "login";
     echo "<script>alert('no existe un inicio de secion');window.location='/base-de-datos-escuela/'</script>";
 }
 ?>
